+++ lib/mgmt.js
@@ -9,6 +9,7 @@ const fxNet  = require('fxNetSocket').NetCoonection;
 const daemon = require('fxNetSocket').daemon;
 const parser = require('fxNetSocket').parser;
 const NSLog  = require('fxNetSocket').logger.getInstance();
+const edittor= require('./AssignEdittor.js');
 const events = require('events');
 const fs     = require('fs');
 const util   = require('util');
@@ -63,6 +64,7 @@ function mgmt(delegate, cfg, port) {
     this.loopSaveHeapUsed(saveHeapTime);*/
     // this.getAssign();
     // setTimeout(this.unitTest.bind(this), 5000);
+    this.AssignPath = AssignPath;
 }
 mgmt.prototype.unitTest = function () {
 
@@ -381,7 +383,12 @@ mgmt.prototype.outOfRangeMemLimit = function (assign) {
     server = null;
     group = null;
 };
-/** 自動檢查機制 **/
+/**
+ * 自動檢查機制 start auto remove mem
+ * @param arr_assign {array} assign list
+ * @param sec {number} clock time
+ * @return {Object|number}
+ */
 mgmt.prototype.automaticCheckCluster = function (arr_assign, sec) {
     var self = this;
 
@@ -524,114 +531,19 @@ mgmt.prototype.stopClusterInfos = function () {
 };
 /** 09.26 - get child process list **/
 mgmt.prototype.getAssign    = function (client) {
-    var assign = require('fxNetSocket').getConfig(AssignPath);
-
-    client.write(JSON.stringify({
-        "event": "getAssign",
-        "data" : assign
-    }));
-
+    edittor.getAssign.apply(this, arguments);
 };
 /** write fork object to config file **/
 mgmt.prototype.updateAssign = function (obj, client) {
-
-    var path = AssignPath;
-
-    try {
-        if (typeof obj == "undefined" || typeof obj != "object") {
-            throw new Error("object is not NULL.");
-        }
-        if (typeof obj["assign"] == "undefined" || typeof obj["assign"] != "string") {
-            throw new Error("Cannot read property 'assign' of undefined");
-        }
-        if (typeof obj["file"] == "undefined" || typeof obj["file"] != "string") {
-            throw new Error("Cannot read property 'file' of undefined");
-        }
-        var data = fs.readFileSync(path);
-        var conf = eval("("+data+")");
-        conf["cluster"].push(obj);
-        fs.writeFileSync(path, JSON.stringify(conf));
-        if (typeof client != "undefined")
-            this._writeException(client,this.ADMIN_EVENT_TYPE.SUCCESSFUL, "updateAssgin");
-    }
-    catch (e) {
-        console.log('Configuation load file error:', e);
-        if (typeof client != "undefined")
-            this._writeException(client,this.ADMIN_EVENT_TYPE.INVALID_ARGUMENT);
-    }
+    edittor.updateAssign.apply(this, arguments);
 };
 
 mgmt.prototype.editAssign   = function (oAssign, obj, client) {
-
-    var path = AssignPath;
-
-    try {
-        if (typeof obj == "undefined" || typeof obj != "object") {
-            throw new Error("object is not NULL.");
-        }
-        if (typeof obj["assign"] == "undefined" || typeof obj["assign"] != "string") {
-            throw new Error("Cannot read property 'assign' of undefined");
-        }
-        if (typeof obj["file"] == "undefined" || typeof obj["file"] != "string") {
-            throw new Error("Cannot read property 'file' of undefined");
-        }
-        if (typeof obj["mxoss"] == "undefined" || typeof parseInt(obj["mxoss"]) != "number") {
-            obj["mxoss"] = 1024;
-
-        }
-
-        var data = fs.readFileSync(path);
-        var conf = eval("("+data+")");
-
-        for (var i = 0; i < conf["cluster"].length; i++) {
-
-            if (conf["cluster"][i]["assign"] == oAssign) {
-                conf["cluster"][i]["assign"] = obj["assign"];
-                conf["cluster"][i]["mxoss"] = parseInt(obj["mxoss"]);
-                conf["cluster"][i]["file"] = obj["file"];
-                break;
-            }
-        }
-        console.log('writeFileSync+++++++++',obj);
-        fs.writeFileSync(path, JSON.stringify(conf));
-
-        if (typeof client != "undefined")
-            this._writeException(client,this.ADMIN_EVENT_TYPE.SUCCESSFUL, "editAssgin");
-    }
-    catch (e) {
-        console.log('Configuation load file error:', e);
-        if (typeof client != "undefined")
-            this._writeException(client,this.ADMIN_EVENT_TYPE.INVALID_ARGUMENT);
-    }
+    edittor.editAssign.apply(this, arguments);
 };
 /** delete fork name to config file **/
 mgmt.prototype.deleteAssign = function (name, client) {
-
-    var path = AssignPath;
-
-    try {
-        if (typeof name == "undefined" || typeof name != "string") {
-            throw new Error("Cannot read 'name' of undefined.");
-        }
-        var data = fs.readFileSync(path);
-        var conf = eval("("+data+")");
-
-        for (var i = 0; i < conf.cluster.length; i++) {
-            var obj = conf["cluster"][i];
-            if (obj.assign == name) {
-                delete conf["cluster"][i];
-            }
-        }
-        var saveData = JSON.stringify(conf).replace(",null","");
-        fs.writeFileSync(path, saveData);
-        if (typeof client != "undefined")
-            this._writeException(client,this.ADMIN_EVENT_TYPE.SUCCESSFUL, "deleteAssgin");
-    }
-    catch (e) {
-        console.log('Configuation load file error:', e);
-        if (typeof client != "undefined")
-            this._writeException(client,this.ADMIN_EVENT_TYPE.INVALID_ARGUMENT);
-    }
+    edittor.deleteAssign.apply(this, arguments);
 };
 
 mgmt.prototype.mutexServiceLock = function (bool) {
