+++ lib/remoteSrv.js
@@ -2,15 +2,16 @@
  * Created by Benson.Liao on 2016/9/12.
  */
 
-const net        = require('net');
+const net       = require('net');
 const events    = require('events');
 const util      = require('util');
 const fxNet     = require('fxNetSocket').NetCoonection;
 const NSLog     = require('fxNetSocket').logger.getInstance();
+const edittor   = require('./AssignEdittor.js');
 const sys       = require('./sysstat.js');
-const exec       = require('child_process').exec;
+const exec      = require('child_process').exec;
 const fs        = require('fs');
-const AssPath   = "../../configuration/Assign.json";
+const AssignPath= "../../configuration/Assign.json";
 NSLog.configure({
     logFileEnabled:true,
     consoleEnabled:true,
@@ -112,6 +113,9 @@ remoteContrl.prototype.outside = function () {
                     console.error(e);
                 }
             }
+        });
+        client.manage.on("local", function (json) {
+            client.write(json);
         });
 
         client.heartbeatEnabled = false;
@@ -270,7 +274,7 @@ remoteContrl.prototype.action = function (client, action) {
 
             if (typeof action["data"][0] == "boolean") {
                 client.manage.setLockConnection(action["data"][0]);
-                var path = AssPath;
+                var path = AssignPath;
                 var data = fs.readFileSync(path);
                 var conf = eval("("+data+")");
                 conf["lockState"] = action["data"][0];
@@ -435,6 +439,7 @@ util.inherits(OctoManage, events.EventEmitter); // 繼承事件
 function OctoManage() {
     events.EventEmitter.call(this);
     this.socket = undefined;
+    this.AssignPath = AssignPath;
 }
 OctoManage.prototype.run = function (port) {
     this._init();
@@ -479,6 +484,13 @@ OctoManage.prototype._connect = function (port) {
 OctoManage.prototype.getAssign = function () {
     var sock = this.socket;
     sock.write(JSON.stringify({"event":"getAssign"}));
+    console.log(sock.writable);
+    if (sock.writable) {
+
+    } else {
+        edittor.getAssign.apply(this);
+    }
+
 };
 /** 顯示子程序資訊 **/
 OctoManage.prototype.getClusterInfos = function () {
@@ -522,17 +534,37 @@ OctoManage.prototype.refuseUser = function (assign, bool) {
 /** 新增啟動名單規則 **/
 OctoManage.prototype.updateAssign = function (file, assign, mxoss) {
     var sock = this.socket;
-    sock.write(JSON.stringify({"event":"updateAssign","data":[ {"file":file, "assign":assign, "mxoss":mxoss}]}));
+    var obj  = {"file":file, "assign":assign, "mxoss":mxoss};
+    if (sock.writable) {
+        sock.write(JSON.stringify({"event":"updateAssign","data":[ obj ]}));
+    }else {
+        edittor.updateAssign.apply(this, [obj]);
+        this.getAssign();
+    }
 };
 /** 編輯啟動名單規則 **/
 OctoManage.prototype.editAssign = function (oAssign, file, nAssign, mxoss) {
     var sock = this.socket;
-    sock.write(JSON.stringify({"event":"editAssign","data":[oAssign, {"file":file, "assign":nAssign, "mxoss":mxoss}]}));
+    var obj  = [oAssign, {"file":file, "assign":nAssign, "mxoss":mxoss}];
+
+    if (sock.writable) {
+        sock.write(JSON.stringify({"event":"editAssign","data":obj}));
+    } else {
+        edittor.editAssign.apply(this, obj);
+        this.getAssign();
+    }
+
 };
 /** 刪除指定名稱啟動名單 **/
 OctoManage.prototype.deleteAssign = function (assign) {
     var sock = this.socket;
-    sock.write(JSON.stringify({"event":"deleteAssign","data":[assign]}));
+    if (sock.writable) {
+        sock.write(JSON.stringify({"event":"deleteAssign","data":[assign]}));
+    } else {
+        edittor.deleteAssign.apply(this, [assign]);
+        this.getAssign();
+    }
+
 };
 /** income connect reject or allow **/
 OctoManage.prototype.setLockConnection = function (bool) {
